resume-builder/requirements.txt
flask==2.3.3
pdfminer.six==20221105
docx2txt==0.8
google-generativeai==0.2.0
stqdm==0.0.5
werkzeug==2.3.7
tqdm==4.64.1
